Revise test
- make it so it passes (follow up unittest fixes following unittest CI guidance)
