+++ renthub/tests/test_booking/test_payment_view/test_payment_qr.py
@@ -6,6 +6,8 @@ from django.test import TestCase
 from django.urls import reverse
 from renthub.models import Rental, Room, Renter
 from renthub.utils import Status
+from django.utils import timezone
+from datetime import datetime
 
 
 class PaymentViewTests(TestCase):
@@ -17,44 +19,32 @@ class PaymentViewTests(TestCase):
                                                  phone_number='1234567890')
         self.room1 = Room.objects.create(room_number=101, detail='A cozy room', price=1000.00)
         self.room2 = Room.objects.create(room_number=102, detail='A cozy room', price=1000.00)
-
-        self.rental = Rental.objects.create(room=self.room1, renter=self.renter, price=self.room1.price)
+        self.start_date = timezone.make_aware(datetime(2024, 1, 1))
+        self.end_date = timezone.make_aware(datetime(2025, 1, 1))
 
     def test_qr_base_case(self):
         """A renter can see the QR code if they don't have a Rental for the room."""
         self.client.login(username='renter1', password='testpassword')
-        response = self.client.get(reverse('renthub:payment', kwargs={'room_number': self.room2.room_number}))
+
+        start_date_str = '2024-01'
+        end_date_str = '2024-12'
+        url = reverse('renthub:payment', kwargs={'room_number': self.room2.room_number})
+        url_with_params = f"{url}?start_month={start_date_str}&end_month={end_date_str}"
+
+        response = self.client.get(url_with_params)
         self.assertContains(response, 'Scan this QR code to complete the payment')
 
     def test_qr_not_visible_if_rental_exists(self):
         """A renter cannot see the QR code if they already have a Rental for the room."""
+        Rental.objects.create(room=self.room1, renter=self.renter, price=self.room1.price)
         self.client.login(username='renter1', password='testpassword')
         response = self.client.get(reverse('renthub:payment', kwargs={'room_number': self.room1.room_number}))
         self.assertNotContains(response, 'Scan this QR code to complete the payment')
 
-    def test_qr_not_visible_if_their_latest_rental_request_status_is_wait(self):
-        """A renter cannot see the QR code if they have a RentalRequest for the room and their latest RentalRequest
-        hasn't been rejected yet. """
-        pass
-        # RentalRequest.objects.create(room=self.room2, renter=self.renter, price=1200.00)
-        # self.client.login(username='renter1', password='testpassword')
-        # response = self.client.get(reverse('renthub:payment', kwargs={'room_number': self.room2.room_number}))
-        # self.assertNotContains(response, 'Scan this QR code to complete the payment')
-
-    def test_qr_not_visible_if_their_latest_rental_request_is_approved(self):
-        """A renter cannot see the QR code if they have a RentalRequest for the room and their latest RentalRequest
-        hasn't been rejected yet. """
-        pass
-        # RentalRequest.objects.create(room=self.room2, renter=self.renter, price=1200.00, status=Status.approve)
-        # self.client.login(username='renter1', password='testpassword')
-        # response = self.client.get(reverse('renthub:payment', kwargs={'room_number': self.room2.room_number}))
-        # self.assertNotContains(response, 'Scan this QR code to complete the payment')
-
-    def test_qr_visible_if_their_latest_rental_request_is_rejected(self):
-        """A renter cannot see the QR code if they have a RentalRequest for the room and their latest RentalRequest
+    def test_qr_visible_if_latest_rental_is_rejected(self):
+        """A renter cannot see the QR code if they have a Rental for the room and their latest RentalRequest
         has been rejected. """
-        pass
-        # RentalRequest.objects.create(room=self.room2, renter=self.renter, price=1200.00, status=Status.reject)
-        # self.client.login(username='renter1', password='testpassword')
-        # response = self.client.get(reverse('renthub:payment', kwargs={'room_number': self.room2.room_number}))
-        # self.assertContains(response, 'Scan this QR code to complete the payment')
+        Rental.objects.create(room=self.room2, renter=self.renter, price=1200.00, status=Status.reject)
+        self.client.login(username='renter1', password='testpassword')
+        response = self.client.get(reverse('renthub:payment', kwargs={'room_number': self.room2.room_number}))
+        self.assertContains(response, 'Scan this QR code to complete the payment')
