+++ renthub/tests/test_notification/test_notifications_view.py
@@ -3,7 +3,7 @@ from datetime import timedelta
 from django.test import TestCase
 from django.urls import reverse
 from django.utils import timezone
-from ..models import Notification, Renter
+from renthub.models import Notification, Renter
 
 
 class NotificationViewTest(TestCase):
@@ -17,17 +17,18 @@ class NotificationViewTest(TestCase):
 
         Notification.objects.create(
             renter=self.renter,
-            title='New Notification',
-            message='This is a new notification',
-            post_date=timezone.now(),
-            is_read=False
-        )
-        Notification.objects.create(
-            renter=self.renter,
             title='Old Notification',
             message='This is an old notification',
             post_date=old_notification_date,
             is_read=True
+        )
+
+        Notification.objects.create(
+            renter=self.renter,
+            title='New Notification',
+            message='This is a new notification',
+            post_date=timezone.now(),
+            is_read=False
         )
 
     def test_notification_list_view(self):
@@ -43,7 +44,7 @@ class NotificationViewTest(TestCase):
 
         # Check the order of notifications
         notifications = response.context['notifications']
-        print(notifications)
+        # print(notifications)
         self.assertEqual(notifications[0].title, 'New Notification')  # New should appear first
         self.assertEqual(notifications[1].title, 'Old Notification')
